+++ src/services/episode.ts
@@ -1,9 +1,9 @@
 import { Context, Effect, Layer, Schema } from "effect";
-import { makeRequest } from "@core/client";
+import { makeRequest } from "@transporter";
 import type {
 	MarketOnlyOptions,
 	PaginatedMarketOptions,
-} from "@internal/options";
+} from "@schemas/options";
 import {
 	type GetEpisodeRequest,
 	type GetEpisodeResponse,
@@ -16,16 +16,16 @@ import {
 	GetSavedEpisodeResponseSchema,
 	GetEpisodeResponseSchema,
 	GetSeveralEpisodeResponseSchema,
-} from "@internal/services/episode";
-import type { ApiError } from "@errors/index";
-import type { AuthService } from "auth";
+} from "@schemas/services/episode";
+import type { ApiError } from "@errors";
+import type { AuthService } from "@auth/index";
 import {
 	guardId,
 	guardIds,
 	guardLimit,
 	guardMarket,
 	guardOffset,
-} from "guards";
+} from "@guards";
 
 export class EpisodeService extends Context.Tag("EpisodeService")<
 	EpisodeService,
